+++ options.go
@@ -35,6 +35,9 @@ type Options struct {
 	// Stops on the first failure
 	StopOnFailure bool
 
+	// Fail suite when there are pending or undefined steps
+	Strict bool
+
 	// Forces ansi color stripping
 	NoColors bool